feat: codelist load and use through out the portal DONE
