+++ src/app/modules/activityDetails/utils/formatResults.ts
@@ -2,22 +2,25 @@
 import { ActResultsModel } from '../store/interface';
 import { ListCellModel } from 'app/components/datadisplay/Lists/common/SimpleListItem/model';
 
-/* consts */
-import { resultTypeNames } from 'app/__consts__/iati_standard_code_names';
+/* utils */
+import find from 'lodash/find';
 
 export function formatResults(
-  resultData: ActResultsModel[] | null
+  resultData: ActResultsModel[] | null,
+  resultTypeNames
 ): ListCellModel[][] {
   const results: ListCellModel[][] = [];
   if (resultData) {
     resultData.forEach(result => {
+      const resTypeName = find(resultTypeNames, ['code', result.result_type]);
+
       results.push([
         {
           link: `/result-detail/${encodeURIComponent(result.id)}`,
           value: result.result_title_narrative[0],
         },
         {
-          value: resultTypeNames[result.result_type],
+          value: resTypeName ? resTypeName.name : 'No Data',
         },
         {
           value:
