+++ src/app/modules/activityDetails/utils/formatSingleCardItem.ts
@@ -45,7 +45,8 @@ export function formatSingleCardItem(
         value = 'No Data';
       }
     } else if (element.codeNames) {
-      value = element.codeNames[value];
+      value = find(element.codeNames, ['code', value]);
+      value = value ? value.name : 'No Data';
     }
 
     if (value !== 'No Data' && element.suffix) {
